Restructured frontend folder structure
