+++ frontend/src/pages/Footer/Footer.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Footer = () => {
+  return (
+    <div>
+      hello
+    </div>
+  )
+}
+
+export default Footer
