+++ frontend/src/utils/appStore.js
@@ -1,0 +1,11 @@
+import { configureStore } from '@reduxjs/toolkit'
+import userReducer from './userSlice'
+
+
+const appStore = configureStore({
+  reducer: {
+    user: userReducer,
+  },
+})
+
+export default appStore;
